backend/requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Google AI and Cloud
google-genai==0.2.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-auth==2.27.0

# Data handling
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
pillow==10.2.0

